// products/yappc/frontend/web/src/clients/generated/api/models/ArtifactGraphMergeRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ArtifactGraphMergeRequest = {
    projectId?: string;
    tenantId?: string;
    baseModel?: Record<string, any>;
    leftModel?: Record<string, any>;
    rightModel?: Record<string, any>;
    resolutionStrategy?: string;
};