// products/yappc/frontend/web/src/clients/generated/api/models/ResidualAnalysisRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ResidualIslandRequest } from './ResidualIslandRequest';
export type ResidualAnalysisRequest = {
    projectId?: string;
    tenantId?: string;
    workspaceId?: string;
    residualIslands?: Array<ResidualIslandRequest>;
};